version := "2017.4.2"
Bug fix for taiga issue 248. When a field of type Long is unpopulated, 
it should receive a default value of -1. However this wasn't happening
because default values were being passed as Doubles to the ConvertToLong
function. Fix was to pass "-1" instead of "-1.0" to any case class 
parameter assignment that calls ConvertToLong.

git-svn-id: https://svn.broadinstitute.org/BTL/Parsomatic@1617 49feb11a-dc6b-4521-947f-5d0cd0a85b6f